[release tool] restructure: main.go (#9565)
* rename release/build to release/cmd

build was a folder name for using goyek; cmd is the proper way

* resturture main.go

- create Command interface to return subcommands
- move each subcommand to separate file
- create flags.go to house all flags

* fix release build

* Apply suggestions from code review


* move cmd definitions to own files

---------
